+++ data/swiss/get_napoleon_position.sh
@@ -1,0 +1,6 @@
+# Genome
+https://www.ebi.ac.uk/ena/browser/api/fasta/GCA_003013145.1?download=true&gzip=true
+zcat Napoleon_genome.fa.gz > napoleon_genome.fa
+sed -i "s/ //g" napoleon_genome.fa
+# Mutations
+https://docs.google.com/spreadsheets/d/10SmTGXCfQKU6M8KLqF_lzHDFM4A64pdpVUeNdI7MapA/edit#gid=0