+++ NEWS.md
@@ -1,3 +1,8 @@
+# detectMutation - swiss data 0.2.0.9011
+* README & dag
+* gatk resources
+* napoleon analyses
+
 # detectMutation - swiss data 0.2.0.9010
 * aligning swiss mutation on PM1N: to be cleaned in the main workflow
 * srtelka2tsv & tsv2sql: to be cleaned in the main workflow